# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VSI Vector Store Service",
    "description": "Qdrant-compatible vector store with document upload and processing capabilities. Provides both a web UI and REST API for managing vector collections, uploading documents, and performing semantic search.",
    "version": "1.0.0",
    "contact": {
      "name": "VSI Vector Store Service"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/collections/{collection}/upload": {
      "post": {
        "tags": ["Documents"],
        "summary": "Upload and index a file",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "my_documents"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded and indexed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "filename": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{collection}/create-text": {
      "post": {
        "tags": ["Documents"],
        "summary": "Create and index a text document",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "my_documents"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Meeting Notes"
                  },
                  "content": {
                    "type": "string",
                    "example": "Today we discussed the new product features..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Text document created and indexed successfully"
          }
        }
      }
    },
    "/api/collections/{collection}/search": {
      "post": {
        "tags": ["Search"],
        "summary": "Semantic search in collection",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "my_documents"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "example": "machine learning algorithms"
                  },
                  "limit": {
                    "type": "integer",
                    "default": 10,
                    "example": 20
                  },
                  "filter": {
                    "type": "object",
                    "description": "Qdrant filter object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "query": {
                      "type": "string"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "score": {
                            "type": "number"
                          },
                          "payload": {
                            "type": "object"
                          }
                        }
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{collection}/documents": {
      "get": {
        "tags": ["Documents"],
        "summary": "List documents in collection",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents"
          }
        }
      }
    },
    "/api/collections/{collection}/documents/{id}": {
      "get": {
        "tags": ["Documents"],
        "summary": "Get document by ID",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document details"
          },
          "404": {
            "description": "Document not found"
          }
        }
      },
      "delete": {
        "tags": ["Documents"],
        "summary": "Delete document by ID",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "api-key"
      }
    },
...
  }
}